Blockchain Window added
